Joins browse button: use "AposPiecesManager"
